swiper organization
