+++ src/navigation/MainNavigation.js
@@ -1,31 +1,24 @@
+// import HiprLanding from '../screens/HiprLanding';
+// import Hipr from '../screens/Hipr';
+// import HiprTransactions from '../screens/HiprTransactions';
 
-import React, { Component } from "react";
-import { createStackNavigator, createDrawerNavigator, createSwitchNavigator } from "react-navigation";
 
-// import BlockScanner from "../screens/MainApp/BlockScanner";
-// import Confirm from "../screens/Confirm";
+// import QRCapture from '../screens/QRCapture';
+// import QRCapture2 from '../screens/QRCapture2';
+// import WebViewComponent from '../components/WebViewComponent';
+// import DocumentStorage from '../screens/MainApp/DocumentStorage';
+// import DocumentQRScanner from '../screens/MainApp/DocumentQRScanner';
+import React, { Component } from 'react';
+import { createStackNavigator, createDrawerNavigator, createSwitchNavigator } from 'react-navigation';
 
-// import Hipr from "../screens/Hipr";
-// import HiprTransactions from "../screens/HiprTransactions";
-import Login from "../screens/MainApp/Login";
+import TestSplash from '../screens/MainApp/TestSplash';
 
-// import HiprLanding from "../screens/HiprLanding";
-// import TxSwiperContainer from "../screens/MainApp/TxSwiperContainer";
-// import TxSwiper from "../components/TxSwiper";
-import Wallet from "../screens/MainApp/Wallet";
-import Settings from "../screens/MainApp/Settings";
-// import QRCapture from "../screens/QRCapture";
-// import QRCapture2 from "../screens/QRCapture2";
-// import WebViewComponent from "../components/WebViewComponent";
-// import DocumentStorage from "../screens/MainApp/DocumentStorage";
-// import DocumentQRScanner from "../screens/MainApp/DocumentQRScanner";
-
-import TestSplash from "../screens/MainApp/TestSplash";
-
-import Header from "../components/Headers/Header";
-import SideNavMenu from "../components/SideNavMenu"
-import RegAssetNavigator from "./RegisterAssetNavigation";
-import SupplyChainNavigator from "./SupplyChainNavigation";
+import Login from '../screens/MainApp/Login';
+import Header from '../components/Headers/Header';
+import SideNavMenu from '../components/SideNavMenu'
+import RegAssetNavigator from './RegisterAssetNavigation';
+import SupplyChainNavigator from './SupplyChainNavigation';
+import TrackNavigator from '../features/Track/navigation/TrackNavigation'
 
 
 const MainNavigator = createStackNavigator({
@@ -47,16 +40,13 @@ const MainNavigator = createStackNavigator({
             header: <Header headerTitle={'Supply Chain'} navigation={navigation} />
         })
     },
+    TrackNavigator: {
+        screen: TrackNavigator,
+        navigationOptions: ({ navigation }) => ({
+            header: <Header headerTitle={'Track'} navigation={navigation} />
+        })
+    }
 
-    // Wallet: {
-    //     screen: Wallet,
-    // },
-
-    // BlockScanner: {
-    //     screen: Blockscanner
-    // },
-
-    
 
 },
     {
@@ -79,8 +69,6 @@ const SideMenuNavigator = createDrawerNavigator(
 
     })
 
-
-
 const LoginNav = createSwitchNavigator(
     {
         Login: {
@@ -98,35 +86,3 @@ const LoginNav = createSwitchNavigator(
 export default LoginNav;
 
     
-    // HiprLanding: { screen: HiprLanding },
-    // HiprAssets: { screen: HiprAssets },
-    // HiprTransactions: { screen: HiprTransactions },
-    // Hipr: { screen: Hipr },
-
-    // BlockScanner: { screen: BlockScanner },
-    // TxSwiperContainer: { screen: TxSwiperContainer },
-
-    // TrackAssetList: { screen: TrackAssetList },
-    // TrackAssetOptions: { screen: TrackAssetOptions },
-
-    // SupplyChainAssetList: { screen: SupplyChainAssetList },
-    // SupplyChainTxRx: { screen: SupplyChainTxRx },
-    // SupplyChainReview: { screen: SupplyChainReview },
-    // ImageUpload: { screen: ImageUpload },
-    // Camera: { screen: Camera },
-    // DocUp: { screen: DocUp },
-    // EdiT: { screen: EdiT },
-    // MetricInput: { screen: MetricInput },
-    // Confirm: { screen: Confirm },
-
-    // QRCapture: { screen: QRCapture },
-    // QRCapture2: { screen: QRCapture2 },
-
-    // Wallet: { screen: Wallet },
-    // Settings: { screen: Settings },
-
-    // WebViewComponent: { screen: WebViewComponent },
-    // TxSwiper: { screen: TxSwiper },
-    // DocumentStorage: { screen: DocumentStorage },
-    // DocumentQRScanner: { screen: DocumentQRScanner },
-
